M_18_Django_Cache/mysite/requirements.txt
asgiref==3.5.2
attrs==23.1.0
blinker==1.4
certifi==2023.7.22
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
Django==4.1.6
django-debug-toolbar==4.1.0
django-filter==22.1
djangorestframework==3.14.0
drf-spectacular==0.26.4
httplib2==0.20.2
importlib-metadata==4.6.4
inflection==0.5.1
jeepney==0.7.1
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
Pillow==9.4.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
pytz==2022.7.1
PyYAML==5.4.1
referencing==0.30.0
rpds-py==0.9.2
SecretStorage==3.3.1
sentry-sdk==1.29.2
six==1.16.0
sqlparse==0.4.2
systemd-python==234
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.0.4
wadllib==1.3.6
zipp==1.0.0
